Decoupled StorageRef and reader, deprecated includeStorage
