+++ src/main/scala/com/johnsnowlabs/nlp/embeddings/WordEmbeddings.scala
@@ -31,7 +31,6 @@ class WordEmbeddings(override val uid: String)
       .setStorageRef($(storageRef))
       .setDimension($(dimension))
       .setCaseSensitive($(caseSensitive))
-      .setIncludeStorage($(includeStorage))
 
     model
   }
