+++ src/main/scala/com/johnsnowlabs/storage/StorageHelper.scala
@@ -78,4 +78,3 @@ object StorageHelper {
   }
 
 }
-
